Remove a temporary ssh key 'dle_key_tmp' from the cloud (if exists)
Remove a temporary ssh key 'dle_key_tmp' from the cloud (if exists) before add a new ssh key.

For each new creation of the DLE server, the playbook generates a new unique ssh key and adds this key to the cloud during the deployment of the DLE, and after successful deployment it must be removed from the cloud.

In some situations, for example, if the previous attempt to deploy the server was unsuccessful but the temporary key was added to the cloud and not deleted, then the subsequent launch of the playbook could be unsuccessful due to the fact that the new generated temporary ssh key was not added to the cloud because a key with that name already exists. Because of this, the deployment was terminated due to the lack of access to the server using a public ssh key.